+++ docs/changelog.rst
@@ -6,6 +6,8 @@ Changelog
 Development
 ===========
 - (Fill this out as you fix issues and develop your features).
+- BREAKING CHANGE: Drop support for positional arguments when instantiating a document. #2103
+  - From now on keyword arguments (e.g. `Doc(field_name=value)`) are required.
 
 Changes in 0.18.2
 =================